+++ src/api/Authentication/DatabaseAuthenticationClaimsTransformation.cs
@@ -36,7 +36,7 @@ public class DatabaseAuthenticationClaimsTransformation : IClaimsTransformation
         {
             claimsIdentity.AddClaim(new Claim(ClaimTypes.Role, PolicyNames.Admin));
         }
-        else if (authenticatedUser.IsViewer)
+        else
         {
              claimsIdentity.AddClaim(new Claim(ClaimTypes.Role, PolicyNames.Viewer));
         }
@@ -52,7 +52,7 @@ public class DatabaseAuthenticationClaimsTransformation : IClaimsTransformation
         if (subjectId is null)
             return null;
 
-        var user = dbContext.Users.SingleOrDefault(u => u.SubjectId == subjectId) ?? new User { SubjectId = subjectId, IsViewer = true };
+        var user = dbContext.Users.SingleOrDefault(u => u.SubjectId == subjectId) ?? new User { SubjectId = subjectId };
         user.FirstName = principal.Claims.FirstOrDefault(c => c.Type == ClaimTypes.GivenName)?.Value ?? user.FirstName;
         user.LastName = principal.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Surname)?.Value ?? user.LastName;
         user.Name = $"{user.FirstName[0]}. {user.LastName}";